feat: send discord notif if cas exits
