+++ runner.sh
@@ -8,7 +8,8 @@ if [[ ${STATUS[0]} == "OK" ]]; then
   cd $CAS_PATH && npm run start
   exit_code=$?
   if [[ $exit_code != 0 ]]; then
-    echo "Service exited with non-zero exit code"
+    node ./runner/report-exit.js
+    echo "CAS exited with non-zero exit code"
   fi
 fi
